fix: allow cpu perf events to work on raspi
The cpu perf events instrument the tsc, aperf, and mperf to provide
metrics such as IPUS and measure the running frequency.

On some platforms, such as the raspi, while we have instruction and
cycle counters, we do not have the tsc, aperf, and mperf counters.

This change modifies the perf event sampler making those counters
optional. As a result, this sampler allows us to see instruction
and cycle metrics for platforms that lack the additional counters.