WIP: adds bpf irq accounting
The tick-based approach of accounting cpu_usage sampler by using a
kprobe on `cpuacct_account_field` can result in undercounting of
irq/softirq time.

This PR addresses this by first adding separate metrics for both
the number of hard and soft interrupts as well as the amount of
time spent in the irq handlers.

WIP DO NOT MERGE