add priority system for samplers
Adds a basic priority system for samplers to help control the order
in which samplers execute. This system groups samplers into high,
medium, or low priority. The default priority is medium. Samplers
which are known to be fast with deterministic sample latency can
be marked as high priority. Samplers which are slower or having
variable latency should be marked as low priority.